Python/Pallette/requirements.txt
joblib==1.1.0
numpy==1.22.3
Pillow==9.1.0
scikit-learn==1.1.0
scipy==1.8.0
threadpoolctl==3.1.0
tqdm==4.64.0
